backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-celery-beat==2.7.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
python-ulid==3.0.0
gunicorn==23.0.0
requests==2.32.3

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
factory-boy==3.3.1
responses==0.25.3
